Fix linux / mac build. Review URL: http://codereview.chromium.org/17256
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7711 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/safe_browsing/protocol_parser_unittest.cc
@@ -506,7 +506,7 @@ TEST(SafeBrowsingProtocolParsingTest, TestGetHashWithUnknownList) {
                                   &re_key,
                                   &full_hashes));
 
-  EXPECT_EQ(full_hashes.size(), 1);
+  EXPECT_EQ(full_hashes.size(), static_cast<size_t>(1));
   EXPECT_EQ(memcmp("12345678901234567890123456789012", 
                    &full_hashes[0].hash, sizeof(SBFullHash)), 0);
   EXPECT_EQ(full_hashes[0].list_name, "goog-phish-shavar");
@@ -521,7 +521,7 @@ TEST(SafeBrowsingProtocolParsingTest, TestGetHashWithUnknownList) {
                                   &re_key,
                                   &full_hashes));
 
-  EXPECT_EQ(full_hashes.size(), 2);
+  EXPECT_EQ(full_hashes.size(), static_cast<size_t>(2));
   EXPECT_EQ(memcmp("12345678901234567890123456789012", 
                    &full_hashes[0].hash, sizeof(SBFullHash)), 0);
   EXPECT_EQ(full_hashes[0].list_name, "goog-phish-shavar");
